Update types, add import result to store, use import status to create alert status.
